Release/linter comments (#47)

+++ internal/app/client/movie.go
@@ -14,6 +14,7 @@ import (
 //go:generate mockgen -source=movie.go -destination=../user/service/mocks/movie_mock.go
 //go:generate mockgen -source=movie.go -destination=../movie/service/mocks/mock.go
 
+// MovieClientInterface интерфейс клиента фильмов
 type MovieClientInterface interface {
 	// GetMovieByGenre(ctx context.Context, genre string) ([]models.MovieShortInfo, error)
 	GetMovie(ctx context.Context, mvID int) (*models.MovieInfo, error)
@@ -22,8 +23,8 @@ type MovieClientInterface interface {
 	SearchActors(ctx context.Context, query string) ([]models.ActorInfo, error)
 	GetCollection(ctx context.Context, filter string) ([]models.Collection, error)
 	GetFavorites(ctx context.Context, mvIDs []uint64) ([]models.MovieShortInfo, error)
-	GetUserRating(ctx context.Context, movieID, userID int) (int, error)
-	AddOrUpdateRating(ctx context.Context, movieID, userID, rating int) error
+	GetUserRating(ctx context.Context, movieID, userID int32) (int32, error)
+	AddOrUpdateRating(ctx context.Context, movieID, userID, rating int32) error
 }
 
 // MovieClient struct implements MovieClientInterface
@@ -228,7 +229,7 @@ func (m MovieClient) GetActor(ctx context.Context, actorID int) (*models.ActorIn
 // var respp = make([]models.MovieShortInfo, 0, len(respMovie))
 
 // 	for i, movie := range respMovie {
-// 		respp[i].ID = int(movie.Id)
+// 		respp[i].ID = int(movie.ID)
 // 		respp[i].CardURL = movie.CardUrl
 // 		respp[i].MovieType = movie.MovieType
 // 		respp[i].AlbumURL = movie.AlbumUrl
@@ -332,13 +333,14 @@ func (m MovieClient) GetFavorites(ctx context.Context, mvIDs []uint64) ([]models
 	return ans, nil
 }
 
-func (m MovieClient) GetUserRating(ctx context.Context, movieID, userID int) (int, error) {
+// GetUserRating получение рейтинга
+func (m MovieClient) GetUserRating(ctx context.Context, movieID, userID int32) (int32, error) {
 	start := time.Now()
 	method := "GetUserRating"
 
 	resp, err := m.movieMS.GetUserRating(ctx, &movie.GetUserRatingRequest{
-		MovieId: int32(movieID),
-		UserId:  int32(userID),
+		MovieId: movieID,
+		UserId:  userID,
 	})
 
 	saveMetric(start, movieClient, method, err)
@@ -347,17 +349,18 @@ func (m MovieClient) GetUserRating(ctx context.Context, movieID, userID int) (in
 		return 0, err
 	}
 
-	return int(resp.Rating.Rating), nil
+	return int32(resp.Rating.Rating), nil
 }
 
-func (m MovieClient) AddOrUpdateRating(ctx context.Context, movieID, userID, rating int) error {
+// AddOrUpdateRating добавление рейтинга
+func (m MovieClient) AddOrUpdateRating(ctx context.Context, movieID, userID, rating int32) error {
 	start := time.Now()
 	method := "AddOrUpdateRating"
 
 	_, err := m.movieMS.AddOrUpdateRating(ctx, &movie.AddOrUpdateRatingRequest{
-		MovieId: int32(movieID),
-		UserId:  int32(userID),
-		Rating:  int32(rating),
+		MovieId: movieID,
+		UserId:  userID,
+		Rating:  rating,
 	})
 
 	saveMetric(start, movieClient, method, err)
